% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O.R
\name{a18_c_w}
\alias{a18_c_w}
\title{Calculate 18O/16O fractionation factor between CaCO3 and H2O}
\usage{
a18_c_w(temp, min = "calcite", eq = "Daeron19")
}
\arguments{
\item{temp}{Crystallization temperature, in degrees Celsius.}

\item{min}{Mineralogy. Options are \code{"calcite"} (default), \code{"aragonite"}, and \code{"dolomite"}.}

\item{eq}{Equation used to calculate the equilibrium 18O/16O oxygen isotope
fractionation factor between calcium carbonate and water. Options depend on mineralogy.
For calcite choose from \code{"Daeron19"} (default), \code{"Watkins13"}, \code{"Coplen07"}, \code{"KO97"} , and \code{"FO77"}.
For aragonite and dolomite the eq need not be specified.}
}
\value{
Returns the 18O/16O oxygen isotope fractionation factor "alpha"
and the corresponding isotope fractionation value "epsilon" (parts per mille).
}
\description{
\code{a18_c_w()} calculates the 18O/16O oxygen isotope fractionation factor between calcium carbonate and water
}
\examples{
a18_c_w(25, "calcite")   # Returns 1.030249 and 30.25
a18_c_w(25, "aragonite") # Returns 1.000913 and 0.91
a18_c_w(25, "dolomite")  # Returns 1.031456 and 31.46
}
\references{
\strong{Calcite:}

O'Neil, J. R., Clayton, R. N., & Mayeda, T. K. (1969). Oxygen isotope fractionation in divalent metal carbonates. The Journal of Chemical Physics, 51(12), 5547-5558. \url{https://www.doi.org/10.1063/1.1671982}

Kim, S.-T., & O'Neil, J. R. (1997). Equilibrium and nonequilibrium oxygen isotope effects in synthetic carbonates. Geochimica et Cosmochimica Acta, 61(16), 3461-3475. \url{https://www.doi.org/10.1016/S0016-7037(97)00169-5}

Coplen, T. B. (2007). Calibration of the calcite–water oxygen-isotope geothermometer at Devils Hole, Nevada, a natural laboratory. Geochimica et Cosmochimica Acta, 71(16), 3948-3957. \url{https://www.doi.org/10.1016/j.gca.2007.05.028}

Watkins, J. M., Nielsen, L. C., Ryerson, F. J., & DePaolo, D. J. (2013). The influence of kinetics on the oxygen isotope composition of calcium carbonate. Earth and Planetary Science Letters, 375, 349-360. \url{https://www.doi.org/10.1016/j.epsl.2013.05.054}

Daëron, M., Drysdale, R. N., Peral, M., Huyghe, D., Blamart, D., Coplen, T. B., et al. (2019). Most Earth-surface calcites precipitate out of isotopic equilibrium. Nature Communications, 10, 429. \url{https://www.doi.org/10.1038/s41467-019-08336-5}

\strong{Aragonite:}

Dettman, D. L., Reische, A. K., & Lohmann, K. C. (1999). Controls on the stable isotope composition of seasonal growth bands in aragonitic fresh-water bivalves (unionidae). Geochimica et Cosmochimica Acta, 63(7-8), 1049-1057. \url{https://www.doi.org/10.1016/s0016-7037(99)00020-4}

\strong{Dolomite:}

Vasconcelos, C., McKenzie, J. A., Warthmann, R., & Bernasconi, S. M. (2005). Calibration of the d18O paleothermometer for dolomite precipitated in microbial cultures and natural environments. Geology, 33(4), 317-320. \url{https://www.doi.org/10.1130/g20992.1}
}
\seealso{
\code{\link[=d18O]{d18O()}}
}
