% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{a18_H2O_OH}
\alias{a18_H2O_OH}
\title{Calculate the 18O/16O fractionation factor between water and
hydroxide ion}
\usage{
a18_H2O_OH(temp, eq = "Z21-X3LYP")
}
\arguments{
\item{temp}{Temperature (°C).}

\item{eq}{Equation used to calculate the fractionation factor.
\itemize{
\item \code{Z20-X3LYP}: the X3LYP/6-311+G(d,p) theoretical equation of Zeebe (2020).
\item \code{Z20-MP2}: the MP2/aug-cc-pVDZ theoretical equation of Zeebe (2020).
}}
}
\value{
Returns the 18O/16O oxygen isotope fractionation factor "alpha".
}
\description{
\code{a18_H2O_OH()} calculates the 18O/16O oxygen isotope
fractionation factor between H2O and aqueous hydroxide ion.
}
\examples{
a18_H2O_OH(90, "Z21-X3LYP") # Returns 1.008146

}
\references{
Zeebe, R. E. (2020).
Oxygen isotope fractionation between water and the aqueous hydroxide ion.
Geochimica et Cosmochimica Acta, 289, 182-195.
\url{https://doi.org/10.1016/j.gca.2020.08.025}
}
\seealso{
Other alpha: 
\code{\link{a18_CO2acid_c}()},
\code{\link{a18_c_H2O}()}
}
\concept{alpha}
