% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O.R
\name{alpha18_cc_water}
\alias{alpha18_cc_water}
\title{Calculate 18O/16O fractionation factor}
\usage{
alpha18_cc_water(temperature, eq = "Daeron19")
}
\arguments{
\item{temperature}{Crystallization temperature, in degrees Celsius.}

\item{eq}{Defines the equation used for the calculation. Options are \emph{"KO97", "Coplen07", "Watkins13"}, and \emph{"Daeron19"}.
Default is \emph{"Daeron19"}.}
}
\value{
Returns the 18O/16O oxygen isotope fractionation factor.
}
\description{
\code{alpha18_cc_water()} calculates the d18O value of the ambient water
from the d18O value of a carbonate and its crystallization temperature.
}
\examples{
alpha18_cc_water(33.7,"Coplen07") # Returns -13.54
}
\references{
O'Neil, J. R., Clayton, R. N., & Mayeda, T. K. (1969). Oxygen isotope fractionation in divalent metal carbonates. The Journal of Chemical Physics, 51(12), 5547-5558. \url{https://www.doi.org/10.1063/1.1671982}

Kim, S.-T., & O'Neil, J. R. (1997). Equilibrium and nonequilibrium oxygen isotope effects in synthetic carbonates. Geochimica et Cosmochimica Acta, 61(16), 3461-3475. \url{https://www.doi.org/10.1016/S0016-7037(97)00169-5}

Coplen, T. B. (2007). Calibration of the calcite–water oxygen-isotope geothermometer at Devils Hole, Nevada, a natural laboratory. Geochimica et Cosmochimica Acta, 71(16), 3948-3957. \url{https://www.doi.org/10.1016/j.gca.2007.05.028}

Watkins, J. M., Nielsen, L. C., Ryerson, F. J., & DePaolo, D. J. (2013). The influence of kinetics on the oxygen isotope composition of calcium carbonate. Earth and Planetary Science Letters, 375, 349-360. \url{https://www.doi.org/10.1016/j.epsl.2013.05.054}

Daëron, M., Drysdale, R. N., Peral, M., Huyghe, D., Blamart, D., Coplen, T. B., et al. (2019). Most Earth-surface calcites precipitate out of isotopic equilibrium. Nature Communications, 10, 429. \url{https://www.doi.org/10.1038/s41467-019-08336-5}
}
\seealso{
\code{\link[=d18O]{d18O()}}
}
