% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d17O.R
\name{d17Oc}
\alias{d17Oc}
\title{Triple oxygen isotope values}
\usage{
d17Oc(temp, d18Ow_VSMOW, eq18 = "Daeron19", lambda = 0.528)
}
\arguments{
\item{temp}{Calcite growth temperature (°C).}

\item{d18Ow_VSMOW}{Water d18O value expressed on the VSMOW scale (‰).}

\item{eq18}{Equation used to calculate the equilibrium 18O/16O oxygen isotope
fractionation factor between calcite and water.
Options are \code{"Daeron19"} (default), \code{"Watkins13"}, \code{"Coplen07"}, \code{"KO97"} , and \code{"FO77"}.}

\item{lambda}{Triple oxygen isotope reference slope. Default is \code{0.528}.}
}
\value{
Returns a data frame:
\itemize{
\item d18O value of the carbonate expressed on the VSMOW scale (‰).
\item d18O value of the carbonate expressed on the VSMOW scale (‰).
\item D17O value of the carbonate expressed on the VSMOW scale (‰).
}
}
\description{
\code{d17Oc()} calculates equilibrium calcite d18O, d17O, and D17O values for a given temperature.
}
\details{
\deqn{\theta_{A/B} = \frac{\alpha^{17}_{A/B}}{\alpha^{18}_{A/B}}}

\deqn{ \delta'^{17}O_{w,VSMOW} = \beta \times \delta'^{18}O_{w,VSMOW} + \gamma \textrm{ where } \beta=0.528 \textrm{ and } \gamma = 0 }

\deqn{\Delta^{17}O = \delta'^{17}O_{c,VSMOW} - \lambda \times \delta'^{18}O_{c,VSMOW} }
}
\examples{
d17Oc(10,-1) # Returns d18Oc = 32.44, d17Oc = 16.91, D17O = -0.084
d17Oc(10,-1)[,3] # Returns D17O = -0.084
prime(d17Oc(10,-1)[,2]) # Returns d'17O = 16.77

}
\references{
Guo, W., & Zhou, C. (2019).
Triple oxygen isotope fractionation in the DIC-H2O-CO2 system: A numerical framework and its implications.
Geochimica et Cosmochimica Acta, 246, 541-564.
\url{https://www.doi.org/10.1016/j.gca.2018.11.018}
}
\seealso{
Other equilibrium_carbonate: 
\code{\link{D47c}()},
\code{\link{D48c}()},
\code{\link{d18Oc}()}
}
\concept{equilibrium_carbonate}
