% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O.R
\name{d18Oc}
\alias{d18Oc}
\title{Equilibrium carbonate d18O value}
\usage{
d18Oc(temp, d18Ow_VSMOW, min = "calcite", eq = "Daeron19")
}
\arguments{
\item{temp}{Crystallization temperature in degrees Celsius.}

\item{d18Ow_VSMOW}{Water d18O value expressed on the VSMOW scale (parts per mille).}

\item{min}{Mineralogy. Options are \code{"calcite"} (default), \code{"aragonite"}, and \code{"dolomite"}.}

\item{eq}{Equation used to calculate the equilibrium 18O/16O oxygen isotope
fractionation factor between calcium carbonate and water. Options depend on mineralogy.
For calcite choose from \code{"Daeron19"} (default), \code{"Watkins13"}, \code{"Coplen07"}, \code{"KO97"} , and \code{"FO77"}.
For aragonite and dolomite the eq need not be specified.}
}
\value{
Returns the equilibrium carbonate d18O value expressed on the VSMOW scale (parts per mille).
}
\description{
\code{d18Oc()} calculates the equilibrium d18O value of a carbonate grown
at a given temperature.
}
\note{
Use \code{\link[=to_vsmow]{to_vsmow()}} and \code{\link[=to_vpdb]{to_vpdb()}} to convert between the VSMOW and VPDB scales.
}
\examples{
d18Oc(33.7, -13.54, eq="Coplen07") # Returns 14.58
to_vpdb(d18Oc(12, -6.94, min="aragonite")) # Returns -5.21
d18Oc(25, -10.96, min="dolomite") # Returns 20.15

}
\references{
References are listed in the description of \code{\link[=a18c_w]{a18c_w()}}.
}
\seealso{
\code{\link[=d18Ow]{d18Ow()}} calculates the d18O value of the ambient water
from the d18O value of a carbonate and its growth temperature.

Other equilibrium_carbonate: 
\code{\link{D47c}()},
\code{\link{D48c}()},
\code{\link{d17Oc}()}
}
\concept{equilibrium_carbonate}
