% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D47.R
\name{D47}
\alias{D47}
\title{Carbonate D47 for given temperatures}
\usage{
D47(temp, eq = "Fiebig21")
}
\arguments{
\item{temp}{Carbonate growth temperature in degrees Celsius.}

\item{eq}{Equation used for the calculation. Options are \code{"Fiebig21"} (default) and \code{"Petersen19"}.
\code{"Fiebig21"} refers to the CDES90 calibration in Fiebig et al. (2021).
\code{"Petersen19"} refers to the synthetic-only D47-RFACBr,WD "UNICAL" calibration of Petersen et al. (2019).}
}
\value{
Returns carbonate D47 values expressed on the CDES90 scale: referenced to 90°C acid digestion.
}
\description{
\code{D47()} calculates carbonate D47 values for a given temperature.
}
\examples{
D47(33.7) # Returns 0.5713
}
\references{
Petersen, S. V., Defliese, W. F., Saenger, C., Daëron, M., Huntington, K. W., John, C. M., et al. (2019). Effects of improved 17O correction on interlaboratory agreement in clumped isotope calibrations, estimates of mineral-specific offsets, and temperature dependence of acid digestion fractionation. Geochemistry, Geophysics, Geosystems, 20(7), 3495-3519. \url{https://www.doi.org/10.1029/2018GC008127}

Fiebig, J., Daëron, M., Bernecker, M., Guo, W., Schneider, G., Boch, R., et al. (2021). Calibration of the dual clumped isotope thermometer for carbonates. Geochimica et Cosmochimica Acta. \url{https://www.doi.org/10.1016/j.gca.2021.07.012}
}
\seealso{
\code{\link[=temp_D47]{temp_D47()}}

\code{\link[=D48]{D48()}}
}
