% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LR04}
\alias{LR04}
\title{A Pliocene-Pleistocene benthic foraminifera d18O stack}
\format{
A data frame with 2115 rows and 3 variables:
\describe{
\item{age}{Age of the sample expressed as thousands of years
before present (ka).}
\item{d18O_VPDB}{Carbonate d18O value expressed on the VPDB scale (‰).}
\item{d18O_error}{Standard error on the d18O value.}
}
}
\source{
\url{https://lorraine-lisiecki.com/stack.html}
}
\usage{
LR04
}
\description{
A dataset containing the LR04 benthic d18O stack.
}
\references{
Lisiecki, L. E., & Raymo, M. E. (2005).
A Pliocene-Pleistocene stack of 57 globally distributed
benthic d18O records.
Paleoceanography, 20(1), PA1003.
\doi{10.1029/2004pa001071}
}
\seealso{
Other "datasets": 
\code{\link{GTS2020}},
\code{\link{devilshole}},
\code{\link{meteoric_water}}
}
\concept{"datasets"}
\keyword{datasets}
