% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteoric_water}
\alias{meteoric_water}
\title{Oxygen isotope values for meteoric waters}
\format{
A data frame with 156 rows and 4 variables:
\describe{
\item{Sample}{Sample ID as in the original publication.}
\item{d17O}{Water d17O value expressed on the VSMOW scale (‰).}
\item{d18O}{Water d18O value expressed on the VSMOW scale (‰).}
\item{Reference}{Abbreviated reference for the data point.}
}
}
\usage{
meteoric_water
}
\description{
A dataset containing a compilation of d17O and d17O values
of various meteoric waters.
}
\references{
Luz, B., & Barkan, E. (2010).
Variations of 17O/16O and 18O/16O in meteoric waters.
Geochimica et Cosmochimica Acta, 74(22), 6276–6286.
\doi{10.1016/j.gca.2010.08.016}

Aron, P. G., Levin, N. E., Beverly, E. J., Huth, T. E.,
Passey, B. H., Pelletier, E. M., Poulsen, C. J.,
Winkelstern, I. Z., & Yarian, D. A. (2021).
Triple oxygen isotopes in the water cycle.
Chemical Geology, 565, 116770.
\doi{10.1016/j.chemgeo.2020.120026}
}
\seealso{
Other "datasets": 
\code{\link{GTS2020}},
\code{\link{LR04}},
\code{\link{devilshole}}
}
\concept{"datasets"}
\keyword{datasets}
