% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime.R
\name{prime}
\alias{prime}
\title{Converting delta to delta prime}
\usage{
prime(classical)
}
\arguments{
\item{classical}{"Classical delta" values to be converted (parts per mille).}
}
\value{
Returns the "delta prime" value (parts per mille).
}
\description{
\code{prime()} converts "classical delta" values to "delta prime"" values
}
\details{
\deqn{\delta'^{17}O = 1000 \times \ln(\frac{\delta^{17}O}{1000}+1)}
}
\examples{
prime(10) # Returns 9.950331
}
\seealso{
\code{\link[=unprime]{unprime()}}
}
