% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D47.R
\name{temp_D47}
\alias{temp_D47}
\title{Clumped isotope thermometry}
\usage{
temp_D47(D47_CDES90, D47_error, eq = "Petersen19")
}
\arguments{
\item{D47_CDES90}{Carbonate D47 values expressed on the CDES90 scale: referenced to 90°C acid digestion.}

\item{D47_error}{Error on the D47 value. Optional.}

\item{eq}{Equation used for the calculation. Options are \code{"Petersen19"} (default) and \code{"Kele15"}.
\code{"Petersen19"} refers to the synthetic-only IUPAC-reprocessed "Br "UNICAL" calibration of Petersen et al. (2019).
\code{"Kele14"} refers to the Kele et al. (2015) calibration reprocessed by Bernasconi et al. (2020) using the IUPAC parameters.}
}
\value{
Returns the carbonate growth temperature in degrees Celsius,
and — if \code{D47_error} is specified — also the error.
}
\description{
\code{temp_D47()} calculates carbonate growth temperature from D47 value.
}
\examples{
temp_D47(0.580) # Returns 33.7
temp_D47(0.580, 0.004) # Returns 33.7 and 1.9

}
\references{
References are listed at \code{\link[=D47c]{D47c()}}.
}
\seealso{
\code{\link[=D47c]{D47c()}} calculates the equilibrium carbonate D47 value for a given temperature.

Other thermometry: 
\code{\link{temp_D48}()},
\code{\link{temp_d18O}()}
}
\concept{thermometry}
