% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpdb.R
\name{to_vpdb}
\alias{to_vpdb}
\title{Converting between VSMOW and VPDB scales}
\usage{
to_vpdb(d18O_VSMOW, eq = "IUPAC")
}
\arguments{
\item{d18O_VSMOW}{d18O values expressed on the VSMOW scale (parts per mille).}

\item{eq}{Equation used for the conversion. Options are \code{"IUPAC"}, and \code{"Coplen83"}.
The default is \code{"IUPAC"}: the IUPAC recommended equation listed in Brand et al. (2014) and Kim et al. (2015).
To use the equation listed in Coplen et al. (1983) and the Hoefs book, set the parameter to \code{"Coplen83"}.}
}
\value{
Returns the d18O value expressed on the VPDB scale (parts per mille).
}
\description{
\code{to_vpdb()} converts d18O values expressed on the VSMOW scale to the VPDB scale.
}
\details{
The IUPAC recommended equation to convert between the scales is:

\deqn{\delta^{18}O_{VPDB} = 0.97001 \times \delta^{18}O_{VSMOW} - 29.99 }
}
\examples{
to_vpdb(0) # Returns -29.99
to_vpdb(0, eq = "Coplen83") # Returns -29.98

}
\references{
References are listed at \code{\link[=to_vsmow]{to_vsmow()}}.
}
\seealso{
\code{\link[=to_vsmow]{to_vsmow()}} converts d18O values expressed on the VPDB scale to the VSMOW scale.
}
