% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime.R
\name{unprime}
\alias{unprime}
\title{Converting d'18O to d18O}
\usage{
unprime(d18O_prime)
}
\arguments{
\item{d18O_prime}{d18O values to be converted (parts per mille).}
}
\value{
Returns the d18O value (parts per mille).
}
\description{
\code{unprime()} converts d'18O values to d18O.
}
\examples{
unprime(9.950331) # Returns 10
}
\seealso{
\code{\link[=prime]{prime()}}
}
