% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpdb.R
\name{VSMOW}
\alias{VSMOW}
\title{Converting between VPDB and VSMOW scales}
\usage{
VSMOW(d18O_VPDB, eq = "IUPAC")
}
\arguments{
\item{d18O_VPDB}{d18O values expressed on the VPDB scale (parts per mille).}

\item{eq}{Defines the equation used for the conversion. Options are \emph{"IUPAC"}, and \emph{"Coplen83"}.
The default is \emph{"IUPAC"}: the IUPAC recommended equation listed in Brand et al. (2014) and Kim et al. (2015).
To use the equation listed in Coplen et al. (1983) and the Hoefs book, set the parameter to \emph{"Coplen83"}.}
}
\value{
Returns d18O values expressed on the VSMOW scale (parts per mille).
}
\description{
\code{VMSOW()} converts d18O values expressed on the VPDB scale to the VSMOW scale.
}
\examples{
VSMOW(0) # Returns 30.92
VSMOW(0, eq="Coplen83") # Returns 30.91
}
\references{
Coplen, T. B., Kendall, C., & Hopple, J. (1983). Comparison of stable isotope reference samples. Nature, 302, 236-238. \url{https://www.doi.org/10.1038/302236a0}

Brand, W. A., Coplen, T. B., Vogl, J., Rosner, M., & Prohaska, T. (2014). Assessment of international reference materials for isotope-ratio analysis (IUPAC Technical Report). Pure and Applied Chemistry, 86(3), 425-467. \url{https://www.doi.org/10.1515/pac-2013-1023}

Kim, S.-T., Coplen, T. B., & Horita, J. (2015). Normalization of stable isotope data for carbonate minerals: Implementation of IUPAC guidelines. Geochimica et Cosmochimica Acta, 158, 276-289. \url{https://www.doi.org/10.1016/j.gca.2015.02.011}
}
\seealso{
\code{\link[=VPDB]{VPDB()}}
}
