% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic.R
\name{xdic}
\alias{xdic}
\title{Dissolved inorganic carbon species}
\usage{
xdic(temp, pH, S)
}
\arguments{
\item{temp}{The temperature of the solution in degrees Celsius.}

\item{pH}{The pH of the solution.}

\item{S}{The salinity of the solution in parts per mille.}
}
\value{
Returns a data frame with the relative abundance of the DIC species.
}
\description{
\code{xdic()} calculates the relative abundance of the DIC species
as a function of solution temperature, pH, and salinity.
}
