% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O.R
\name{a18_c_H2O}
\alias{a18_c_H2O}
\title{Calculate the 18O/16O fractionation factor between carbonate and water}
\usage{
a18_c_H2O(temp, min, eq)
}
\arguments{
\item{temp}{Crystallization temperature in degrees Celsius.}

\item{min}{Mineralogy. Options are \code{"calcite"}, \code{"aragonite"},
and \code{"dolomite"}.}

\item{eq}{Equation used for the calculations.

Options for \strong{calcite}:
\itemize{
\item \code{"Daeron19"}: the equation of Daëron et al. (2019)
\item \code{"Watkins13"}: the equation of Watkins et al. (2013)
\item \code{"Coplen07"}: the equation of Coplen (2007)
\item \code{"KO97"}: Kim and O'Neil (1997)
\item \code{"FO77"}: the equation of O'Neil et al. (1969)
reprocessed by Friedman and O'Neil (1977)
}

Options for \strong{aragonite}:
\itemize{
\item \code{"Dettman99"}: the equation of Grossman and Ku (1986)
reprocessed by Dettman et al. (1999)
}

Options for \strong{dolomite}:
\itemize{
\item \code{"Vasconcelos05"}: the equation of Vasconcelos et al. (2005)
}}
}
\value{
Returns the equilibrium 18O/16O fractionation factor "alpha".
}
\description{
\code{a18_c_H2O()} calculates the equilibrium 18O/16O fractionation factor
between carbonate and water for a given temperature.
}
\examples{
a18_c_H2O(25, "calcite", "Coplen07")
a18_c_H2O(25, "aragonite", "Dettman99")
a18_c_H2O(25, "dolomite", "Vasconcelos05")

}
\references{
\strong{Calcite:}

O'Neil, J. R., Clayton, R. N., & Mayeda, T. K. (1969).
Oxygen isotope fractionation in divalent metal carbonates.
The Journal of Chemical Physics, 51(12), 5547-5558.
\url{https://doi.org/10.1063/1.1671982}

Kim, S.-T., & O'Neil, J. R. (1997).
Equilibrium and nonequilibrium oxygen isotope effects
in synthetic carbonates.
Geochimica et Cosmochimica Acta, 61(16), 3461-3475.
\url{https://doi.org/10.1016/S0016-7037(97)00169-5}

Coplen, T. B. (2007).
Calibration of the calcite–water oxygen-isotope geothermometer
at Devils Hole, Nevada, a natural laboratory.
Geochimica et Cosmochimica Acta, 71(16), 3948-3957.
\url{https://doi.org/10.1016/j.gca.2007.05.028}

Watkins, J. M., Nielsen, L. C., Ryerson, F. J., & DePaolo, D. J. (2013).
The influence of kinetics on the oxygen isotope composition
of calcium carbonate.
Earth and Planetary Science Letters, 375, 349-360.
\url{https://doi.org/10.1016/j.epsl.2013.05.054}

Daëron, M., Drysdale, R. N., Peral, M., Huyghe, D., Blamart, D.,
Coplen, T. B., et al. (2019).
Most Earth-surface calcites precipitate out of isotopic equilibrium.
Nature Communications, 10, 429.
\url{https://doi.org/10.1038/s41467-019-08336-5}

\strong{Aragonite:}

Dettman, D. L., Reische, A. K., & Lohmann, K. C. (1999).
Controls on the stable isotope composition of seasonal growth bands
in aragonitic fresh-water bivalves (unionidae).
Geochimica et Cosmochimica Acta, 63(7-8), 1049-1057.
\url{https://doi.org/10.1016/s0016-7037(99)00020-4}

\strong{Dolomite:}

Vasconcelos, C., McKenzie, J. A., Warthmann, R.,
& Bernasconi, S. M. (2005).
Calibration of the d18O paleothermometer for dolomite precipitated in
microbial cultures and natural environments.
Geology, 33(4), 317-320.
\url{https://doi.org/10.1130/g20992.1}
}
\seealso{
Other fractionation_factors: 
\code{\link{a18_CO2acid_c}()},
\code{\link{a18_H2O_OH}()},
\code{\link{a_A_B}()}
}
\concept{fractionation_factors}
