% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{a18_CO2acid_c}
\alias{a18_CO2acid_c}
\title{Acid fractionation factor: 18O/16O}
\usage{
a18_CO2acid_c(temp, min)
}
\arguments{
\item{temp}{Acid digestion temperature (°C).}

\item{min}{Mineralogy. Options are \code{"calcite"} and \code{"aragonite"}.}
}
\value{
Returns the 18O/16O oxygen isotope fractionation factor "alpha".
}
\description{
\code{a18_CO2acid_c()} calculates the 18O/16O oxygen isotope
fractionation factor between CO2 from acid digestion and carbonate.
}
\examples{
a18_CO2acid_c(90, "calcite") # Returns 1.008146

}
\references{
Sharma, T., & Clayton, R. N. (1965).
Measurement of ratios of total oxygen of carbonates.
Geochimica et Cosmochimica Acta, 29(12), 1347-1353.
\url{https://doi.org/10.1016/0016-7037(65)90011-6}

Kim, S.-T., Mucci, A., & Taylor, B. E. (2007).
Phosphoric acid fractionation factors for calcite and aragonite
between 25 and 75 °C: Revisited.
Chemical Geology, 246(3-4), 135-146.
\url{https://doi.org/10.1016/j.chemgeo.2007.08.005}

Kim, S.-T., Coplen, T. B., & Horita, J. (2015).
Normalization of stable isotope data for carbonate minerals:
Implementation of IUPAC guidelines.
Geochimica et Cosmochimica Acta, 158, 276-289.
\url{https://doi.org/10.1016/j.gca.2015.02.011}
}
\seealso{
Other alpha: 
\code{\link{a18_H2O_OH}()},
\code{\link{a18_c_H2O}()}
}
\concept{alpha}
