% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{A_from_a}
\alias{A_from_a}
\title{Isotope composition from fractionation factor}
\usage{
A_from_a(a, B)
}
\arguments{
\item{a}{Isotope fractionation factor between A/B "alpha".}

\item{B}{Delta value of B (‰).}
}
\value{
Returns the delta value of B (‰).
}
\description{
\code{A_from_a()} calculates the delta value of A from the
isotope fractionation factor and the delta value of B.
}
\examples{
A_from_a(1.033, -10)

}
\seealso{
\code{\link[=a_A_B]{a_A_B()}} calculates the isotope fractionation factor
between A and B.

\code{\link[=A_from_a]{A_from_a()}} calculates the delta value of B.
}
