% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{B_from_a}
\alias{B_from_a}
\title{Isotope composition from fractionation factor}
\usage{
B_from_a(a, A)
}
\arguments{
\item{a}{Isotope fractionation factor between A/B.}

\item{A}{Delta value of A (‰).}
}
\value{
Returns the delta value of B (‰).
}
\description{
\code{B_from_a()} calculates the delta value of B using
an isotope fractionation factor alpha.
}
\examples{
B_from_a(1.033, 12)

}
\seealso{
\code{\link[=a_A_B]{a_A_B()}} calculates the isotope fractionation
factor between A and B.

\code{\link[=A_from_a]{A_from_a()}} calculates the delta value of B.
}
