% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d17O.R
\name{d17O}
\alias{d17O}
\title{Triple oxygen isotope values}
\usage{
d17O(temp, d18Ow_VSMOW, eq18 = "Daeron19", lambda = 0.528)
}
\arguments{
\item{temp}{Calcite growth temperature in degrees Celsius.}

\item{d18Ow_VSMOW}{Water d18O value expressed on the VSMOW scale (parts per mille).}

\item{eq18}{Equation used to calculate the equilibrium 18O/16O oxygen isotope
fractionation factor between calcite and water.
Options are \code{"Daeron19"} (default), \code{"Watkins13"}, \code{"Coplen07"}, \code{"KO97"} , and \code{"FO77"}.}

\item{lambda}{Triple oxygen isotope reference slope. Default is \code{0.528}.}
}
\value{
Returns a data frame with the carbonate d18O, d17O, and D17O values expressed on the VSMOW scale (all in parts per mille).
}
\description{
\code{d17O()} calculates calcite d18O, d17O, and D17O values for given temperatures.
}
\details{
\deqn{\theta_{A/B} = \frac{\alpha^{17}_{A/B}}{\alpha^{18}_{A/B}}}

\deqn{ \delta'^{17}O_{w,VSMOW} = \beta \times \delta'^{18}O_{w,VSMOW} + \gamma \textrm{ where } \beta=0.528 \textrm{ and } \gamma = 0 }

\deqn{\Delta^{17}O = \delta'^{17}O_{c,VSMOW} - \lambda \times \delta'^{18}O_{c,VSMOW} }
}
\examples{
d17O(10,-1) # Returns d18Oc = 32.44, d17Oc = 16.91, D17O = -0.084
d17O(10,-1)[,3] # Returns D17O = -0.084
prime(d17O(10,-1)[,2]) # Returns d'17O = 16.77
}
\references{
Guo, W., & Zhou, C. (2019). Triple oxygen isotope fractionation in the DIC-H2O-CO2 system: A numerical framework and its implications. Geochimica et Cosmochimica Acta, 246, 541-564. \url{https://www.doi.org/10.1016/j.gca.2018.11.018}

Daëron, M., Drysdale, R. N., Peral, M., Huyghe, D., Blamart, D., Coplen, T. B., et al. (2019). Most Earth-surface calcites precipitate out of isotopic equilibrium. Nature Communications, 10, 429. \url{https://www.doi.org/10.1038/s41467-019-08336-5}
}
\seealso{
\code{\link[=d18O]{d18O()}}
}
