% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d17O.R
\name{d17O_c}
\alias{d17O_c}
\title{Triple oxygen isotope values}
\usage{
d17O_c(temp, d18O_H2O_VSMOW, eq18 = "Daeron19", lambda = 0.528)
}
\arguments{
\item{temp}{Calcite growth temperature (°C).}

\item{d18O_H2O_VSMOW}{Water d18O value expressed on the VSMOW scale (‰).}

\item{eq18}{Equation used to calculate the equilibrium 18O/16O
fractionation factor between calcite and water.
Options are as in \code{\link[=a18_c_H2O]{a18_c_H2O()}} with \code{"Daeron19"} being here the default.}

\item{lambda}{Triple oxygen isotope reference slope. Default is \code{0.528}.}
}
\value{
Returns a data frame:
\itemize{
\item d18O value of the carbonate expressed on the VSMOW scale (‰).
\item d18O value of the carbonate expressed on the VSMOW scale (‰).
\item D17O value of the carbonate expressed on the VSMOW scale (‰).
}
}
\description{
\code{d17O_c()} calculates equilibrium calcite d18O, d17O, and D17O values
for a given temperature.
}
\details{
\deqn{\theta_{A/B} = \frac{\alpha^{17}_{A/B}}{\alpha^{18}_{A/B}}}

\deqn{ \delta'^{17}O_{w,VSMOW} =
\beta \times \delta'^{18}O_{w,VSMOW} + \gamma
\textrm{ where } \beta=0.528 \textrm{ and } \gamma = 0 }

\deqn{\Delta^{17}O = \delta'^{17}O_{c,VSMOW} -
\lambda \times \delta'^{18}O_{c,VSMOW} }
}
\examples{
d17O_c(10,-1) # Returns the data frame (length = 3)
d17O_c(10,-1)[,1] # Returns the D18O value
prime(d17O_c(10,-1)[,2]) # Returns the d'17O value
d17O_c(10,-1)[,3] # Returns the D17O value

}
\references{
Guo, W., & Zhou, C. (2019).
Triple oxygen isotope fractionation in the DIC-H2O-CO2 system:
A numerical framework and its implications.
Geochimica et Cosmochimica Acta, 246, 541-564.
\url{https://www.doi.org/10.1016/j.gca.2018.11.018}
}
\seealso{
Other equilibrium_carbonate: 
\code{\link{D47}()},
\code{\link{D48}()},
\code{\link{d18O_c}()}
}
\concept{equilibrium_carbonate}
