% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O.R
\name{d18Ow}
\alias{d18Ow}
\title{Calculate water d18O value}
\usage{
d18Ow(temp, d18Oc_VSMOW, min = "calcite", eq = "Daeron19")
}
\arguments{
\item{temp}{Crystallization temperature, in degrees Celsius.}

\item{d18Oc_VSMOW}{Carbonate d18O value expressed on the VSMOW scale (parts per mille).}

\item{min}{Mineralogy. Options are \code{"calcite"} (default), \code{"aragonite"}, and \code{"dolomite"}.}

\item{eq}{Equation used to calculate the equilibrium 18O/16O oxygen isotope
fractionation factor between calcium carbonate and water. Options depend on mineralogy.
For calcite choose from \code{"Daeron19"} (default), \code{"Watkins13"}, \code{"Coplen07"}, \code{"KO97"} , and \code{"FO77"}.
For aragonite and dolomite the eq need not be specified.}
}
\value{
Returns the water d18O value expressed on the VSMOW scale (parts per mille).
}
\description{
\code{d18Ow()} calculates the d18O value of the ambient water.
from the d18O value of a carbonate and its growth temperature.
}
\note{
Use \code{\link[=to_vsmow]{to_vsmow()}} and \code{\link[=to_vpdb]{to_vpdb()}} to convert between the VSMOW and VPDB scales.
}
\examples{
d18Ow(33.7, 14.58, "calcite", "Coplen07") # Returns -13.54
d18Ow(25, to_vsmow(-7.47), "aragonite") # Returns -6.53
d18Ow(25, 20.43, "dolomite") # Returns -10.69

}
\references{
References are listed in the description of \code{\link[=a18c_w]{a18c_w()}}.
}
\seealso{
\code{\link[=d18Oc]{d18Oc()}} calculates the equilibrium d18O value of a carbonate grown at a given temperature.

\code{\link[=temp_d18O]{temp_d18O()}} calculates growth temperatures from oxygen isotope data.
}
