% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumped.R
\name{D47}
\alias{D47}
\title{Carbonate D47 for a given temperature}
\usage{
D47(temp, eq)
}
\arguments{
\item{temp}{Carbonate growth temperature (°C).}

\item{eq}{Equation used for the calculation.
\itemize{
\item \code{"Fiebig21"}: the CDES90 calibration of Fiebig et al. (2021).
\item \code{"Petersen19"}: the synthetic-only IUPAC parameter "UNICAL" calibration
of Petersen et al. (2019).
}}
}
\value{
Returns the carbonate D47 value expressed on the CDES90 scale (‰).
}
\description{
\code{D47()} calculates the equilibrium carbonate D47 value
for a given temperature.
}
\examples{
D47(33.7, "Petersen19")

}
\references{
Petersen, S. V., Defliese, W. F., Saenger, C., Daëron, M., Huntington,
K. W., John, C. M., et al. (2019).
Effects of improved 17O correction on interlaboratory agreement in
clumped isotope calibrations, estimates of mineral-specific offsets,
and temperature dependence of acid digestion fractionation.
Geochemistry, Geophysics, Geosystems, 20(7), 3495-3519.
\url{https://www.doi.org/10.1029/2018GC008127}

Fiebig, J., Daëron, M., Bernecker, M., Guo, W.,
Schneider, G., Boch, R., et al. (2021).
Calibration of the dual clumped isotope thermometer for carbonates.
Geochimica et Cosmochimica Acta.
\url{https://www.doi.org/10.1016/j.gca.2021.07.012}
}
\seealso{
\code{\link[=temp_D47]{temp_D47()}} calculates growth temperature from a D47 value.

Other equilibrium_carbonate: 
\code{\link{D48}()},
\code{\link{d17O_c}()},
\code{\link{d18O_c}()}
}
\concept{equilibrium_carbonate}
