% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D48.R
\name{D48}
\alias{D48}
\title{Carbonate D48 at given temperature}
\usage{
D48(temperature, eq = "Fiebig21")
}
\arguments{
\item{temperature}{Carbonate growth temperature in degrees Celsius.}

\item{eq}{Equation used for the calculation. Options are \code{"Fiebig21"} and \code{"Swart21"}.
Default is \code{"Fiebig21"}, which refers to the CDES90 calibration in Fiebig et al. (2021). \code{"Swart21"} refers to the CDES90 "PBLM1" calibration in Swart et al. (2021).}
}
\value{
Returns carbonate D48 values expressed on the CDES90 scale: referenced to 90°C acid digestion.
}
\description{
\code{D48()} calculates carbonate D48 values for given temperatures.
}
\examples{
D48(33.7) # Returns 0.237
D48(33.7,"Swart21") # Returns 0.239
}
\references{
Fiebig, J., Daëron, M., Bernecker, M., Guo, W., Schneider, G., Boch, R., et al. (2021). Calibration of the dual clumped isotope thermometer for carbonates. Geochimica et Cosmochimica Acta. \url{https://www.doi.org/10.1016/j.gca.2021.07.012}

Swart, P. K., Lu, C., Moore, E., Smith, M., Murray, S. T., & Staudigel, P. T. (2021). A calibration equation between D48 values of carbonate and temperature. Rapid Communications in Mass Spectrometry, 35(17), e9147. \url{https://www.doi.org/10.1002/rcm.9147}
}
\seealso{
\code{\link[=D47]{D47()}}
}
