% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing.R
\name{mix_d17O}
\alias{mix_d17O}
\title{Mixing curves}
\usage{
mix_d17O(d18O_A, d17O_A, d18O_B, d17O_B, lambda = 0.528)
}
\arguments{
\item{d18O_A}{d18O values of component A (parts per mille).}

\item{d17O_A}{d17O values of component A (parts per mille).}

\item{d18O_B}{d18O values of component B (parts per mille).}

\item{d17O_B}{d17O values of component B (parts per mille).}

\item{lambda}{Triple oxygen isotope reference slope. Default is 0.528.}
}
\value{
Returns a data frame that contains the d18O and d17O values of the mixture,
from 100\% A and 0\% B to 0\% A and 100\% B.
}
\description{
Use \code{mix_d17O()} to produce mixing curves in in triple oxygen isotope space (d18O vs. D17O).
}
\examples{
# Mixing between a Mesozoic marine carbonate and a diagentic carbonate
mix_d17O(d17O(10,-1)[1],d17O(10,-1)[2],d17O(100,0)[1],d17O(100,0)[2])
}
\seealso{
\code{\link[=d17O]{d17O()}}
}
