% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d17O.R
\name{mix_d17O}
\alias{mix_d17O}
\title{Mixing curves in triple oxygen isotope space}
\usage{
mix_d17O(d18O_A, d17O_A, d18O_B, d17O_B, lambda = 0.528)
}
\arguments{
\item{d18O_A}{d18O value of component A (parts per mille).}

\item{d17O_A}{d17O value of component A (parts per mille).}

\item{d18O_B}{d18O value of component B (parts per mille).}

\item{d17O_B}{d17O value of component B (parts per mille).}

\item{lambda}{Triple oxygen isotope reference slope. Default \code{0.528}.}
}
\value{
Returns a data frame:
\itemize{
\item d18O value of the mixture at \% mixing
\item d18O value of the mixture \% mixing
\item \% mixing: from 100\% A and 0\% B to 0\% A and 100\% B.
}
}
\description{
\code{mix_d17O()} produces mixing curves in in triple oxygen isotope space (d18O vs. D17O).
}
\examples{
# Mixing between a Mesozoic marine carbonate and a diagentic carbonate
mix_d17O(d17Oc(10,-1)[1],d17Oc(10,-1)[2],d17Oc(100,0)[1],d17Oc(100,0)[2])

}
\seealso{
\code{\link[=d17Oc]{d17Oc()}} calculates equilibrium calcite d18O, d17O, and D17O values for a given temperature.
}
