% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime.R
\name{prime}
\alias{prime}
\title{Converting d18O to d'18O}
\usage{
prime(d18O)
}
\arguments{
\item{d18O}{d18O values to be converted (parts per mille).}
}
\value{
Returns the d'18O value (parts per mille).
}
\description{
\code{prime()} converts d18O values to d'18O
}
\examples{
prime(10) # Returns 9.950331
}
\seealso{
\code{\link[=unprime]{unprime()}}
}
