% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D47.R
\name{temp_D47}
\alias{temp_D47}
\title{Carbonate growth temperature from D47}
\usage{
temp_D47(D47_CDES90, eq = "Petersen19")
}
\arguments{
\item{D47_CDES90}{Carbonate D47 values expressed on the CDES90 scale: referenced to 90°C acid digestion.}

\item{eq}{Equation used for the calculation. Options are \code{"Petersen19"} (default) and \code{"Kele15"}.
\code{"Petersen19"} refers to the synthetic-only IUPAC-reprocessed "Br "UNICAL" calibration of Petersen et al. (2019).
\code{"Kele14"} refers to the Kele et al. (2015) calibration reprocessed by Bernasconi et al. (2020) using the IUPAC parameters.}
}
\value{
Returns the carbonate growth temperature in degrees Celsius.
}
\description{
\code{temp_D47()} calculates carbonate growth temperature from clumped isotope composition (D47).
}
\examples{
temp_D47(0.580) # Returns 33.7
}
\references{
Kele, S., Breitenbach, S. F. M., Capezzuoli, E., Meckler, A. N., Ziegler, M., Millan, I. M., et al. (2015). Temperature dependence of oxygen- and clumped isotope fractionation in carbonates: a study of travertines and tufas in the 6–95 °C temperature range. Geochimica et Cosmochimica Acta, 168, 172-192. \url{https://www.doi.org/10.1016/j.gca.2015.06.032}

Bernasconi, S. M., Müller, I. A., Bergmann, K. D., Breitenbach, S. F. M., Fernandez, A., Hodell, D. A., et al. (2018). Reducing uncertainties in carbonate clumped isotope analysis through consistent carbonate-based standardization. Geochemistry, Geophysics, Geosystems, 19(9), 2895-2914. \url{https://www.doi.org/10.1029/2017gc007385}

Petersen, S. V., Defliese, W. F., Saenger, C., Daëron, M., Huntington, K. W., John, C. M., et al. (2019). Effects of improved 17O correction on interlaboratory agreement in clumped isotope calibrations, estimates of mineral-specific offsets, and temperature dependence of acid digestion fractionation. Geochemistry, Geophysics, Geosystems, 20(7), 3495-3519. \url{https://www.doi.org/10.1029/2018GC008127}
}
\seealso{
\code{\link[=temp_d18O]{temp_d18O()}}

\code{\link[=temp_D48]{temp_D48()}}
}
