% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumped.R
\name{temp_D47}
\alias{temp_D47}
\title{Clumped isotope thermometry}
\usage{
temp_D47(D47_CDES90, D47_error, eq)
}
\arguments{
\item{D47_CDES90}{Carbonate D47 values expressed on the CDES90 scale (‰).}

\item{D47_error}{Error on the D47 value. Optional.}

\item{eq}{Equation used for the calculation.
\itemize{
\item \code{"Petersen19"}: the synthetic-only composite IUPAC-parameter calibration
of Petersen et al. (2019).
\item \code{"Fiebig21"}: the CDES90 calibration of Fiebig et al. (2021), linearized
between 8–80 °C.
}}
}
\value{
Returns the carbonate growth temperature (°C),
and — if \code{D47_error} is specified — also the error.
}
\description{
\code{temp_D47()} calculates carbonate growth temperature from D47 value.
}
\details{
\strong{"Petersen19"}:

\deqn{\Delta_{47, CDES90} =
0.0383 \times \frac{10^{6}}{T^{2}} + 0.170}

\strong{"Fiebig21"}:

\deqn{\Delta_{47, CDES90} =
0.0391 \times \frac{10^{6}}{T^{2}} + 0.1547}
}
\examples{
temp_D47(D47_CDES90 = 0.577, eq = "Petersen19")

}
\references{
References are listed at \code{\link[=D47]{D47()}}.
}
\seealso{
\code{\link[=D47]{D47()}} calculates the equilibrium carbonate D47 value.

Other thermometry: 
\code{\link{temp_D48}()},
\code{\link{temp_d18O}()}
}
\concept{thermometry}
