% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumped.R
\name{temp_D47}
\alias{temp_D47}
\title{Clumped isotope thermometry}
\usage{
temp_D47(D47_CDES90, D47_error, eq)
}
\arguments{
\item{D47_CDES90}{Carbonate D47 values expressed on the CDES90 scale (‰).}

\item{D47_error}{Error on the D47 value. Optional.}

\item{eq}{Equation used for the calculation.
\itemize{
\item \code{"Petersen19"}: the synthetic-only IUPAC parameter "UNICAL" calibration
of Petersen et al. (2019).
\item \code{"Kele14"}: the Kele et al. (2015) calibration
reprocessed by Bernasconi et al. (2020) using the IUPAC parameters.
}}
}
\value{
Returns the carbonate growth temperature (°C),
and — if \code{D47_error} is specified — also the error.
}
\description{
\code{temp_D47()} calculates carbonate growth temperature from D47 value.
}
\examples{
temp_D47(D47_CDES90 = 0.580, eq = "Petersen19")
temp_D47(D47_CDES90 = 0.580, D47_error = 0.004, eq = "Petersen19")

}
\references{
References are listed at \code{\link[=D47]{D47()}}.
}
\seealso{
\code{\link[=D47]{D47()}} calculates the equilibrium carbonate D47 value
for a given temperature.

Other thermometry: 
\code{\link{temp_D48}()},
\code{\link{temp_d18O}()}
}
\concept{thermometry}
