% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D48.R
\name{temp_D48}
\alias{temp_D48}
\title{Dual clumped isotope thermometry}
\usage{
temp_D48(D47_CDES90, D48_CDES90, ks, add = F)
}
\arguments{
\item{D47_CDES90}{Carbonate D47 values expressed on the CDES90 scale.}

\item{D48_CDES90}{Carbonate D48 values expressed on the CDES90 scale.}

\item{ks}{Kinetic slope. Either -0.6 or -1. Has to be negative!}

\item{add}{Add kinetic slope to an already existing plot?}
}
\value{
Returns the carbonate growth temperature in degrees Celsius.
}
\description{
\code{temp_D48()} calculates growth temperatures from D47 and D48 values.
}
\details{
The function calculates a D47 value as an intersect of two curves:
\itemize{
\item the equilibrium D47 vs D48 curve from Fiebig et al. (2021)
\item the kinetic slope
}

The resulting D47 value is then converted to temperature using the \code{\link[=D47]{D47()}} function, i.e., by default the
equation of Petersen et al. (2019). This is not consistent and I will fix it in a later version.
In any case, the resulting discrepancy is smaller than the temperature error.
}
\examples{
temp_D48(0.617, 0.139, -0.6) # Returns 44
temp_D48(0.546, 0.277, -1)   # Returns 33
}
\references{
Fiebig, J., Daëron, M., Bernecker, M., Guo, W., Schneider, G., Boch, R., et al. (2021). Calibration of the dual clumped isotope thermometer for carbonates. Geochimica et Cosmochimica Acta. \url{https://www.doi.org/10.1016/j.gca.2021.07.012}
}
\seealso{
\code{\link[=D48]{D48()}}

\code{\link[=temp_D47]{temp_D47()}}
}
