% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D48.R
\name{temp_D48}
\alias{temp_D48}
\title{Dual clumped isotope thermometry}
\usage{
temp_D48(
  D47_CDES90,
  D48_CDES90,
  D47_error,
  D48_error,
  ks,
  add = FALSE,
  col,
  pch
)
}
\arguments{
\item{D47_CDES90}{Carbonate D47 values expressed on the CDES90 scale (‰).}

\item{D48_CDES90}{Carbonate D48 values expressed on the CDES90 scale (‰).}

\item{D47_error}{Error on the D47 value. Optional.}

\item{D48_error}{Error on the D48 value. Optional.}

\item{ks}{Kinetic slope. Has to be negative!}

\item{add}{Add graphics to an already existing plot? Default \code{FALSE}.}

\item{col}{Graphical parameter. Optional.}

\item{pch}{Graphical parameter. Optional.}
}
\value{
Returns the carbonate growth temperature (‰).
}
\description{
\code{temp_D48()} calculates carbonate growth temperature from D47 and D48 values.
}
\details{
The function calculates a D47 value as an intersect of two curves:
\itemize{
\item the equilibrium D47 vs D48 curve from Fiebig et al. (2021)
\item the kinetic slope
}

The resulting D47 value is then converted to temperature using the \code{\link[=D47c]{D47c()}} function, i.e., by default the
equation of Petersen et al. (2019). This is not consistent and I will fix it in a later version.
In any case, the resulting discrepancy is smaller than the temperature error.
}
\examples{
temp_D48(0.617, 0.139, ks = -0.6) # Returns 44
temp_D48(0.546, 0.277, ks = -1)   # Returns 33

}
\references{
References are listed at \code{\link[=D48c]{D48c()}} and \code{\link[=D47c]{D47c()}}.
}
\seealso{
\code{\link[=D47c]{D47c()}} calculates the equilibrium carbonate D47 value for a given temperature.

\code{\link[=D48c]{D48c()}} calculates the equilibrium carbonate D48 value for a given temperature.

Other thermometry: 
\code{\link{temp_D47}()},
\code{\link{temp_d18O}()}
}
\concept{thermometry}
